refactor(reducer): add more structure to reducers
